a lot of files were changed to make LiSA more std compatible
all deprecated header files were replaced with the std ones
minor problems (name collisions) that resultet from that were solved

-Wno-deprecated was removed and is no longer necessary

+++ src/algorithm/irred/irredtest.cpp
@@ -150,8 +150,8 @@ bool Lisa_IrreducibilityTest::test(Lisa_Graph* plan, Lisa_Graph* comp,const int 
   double count = 0;
   double oldcount = count;
   double tocheck;
-  if(all && param!=GENERATE_SIMILAR) tocheck = pow(2,num-1)-1;
-  else tocheck = pow(2,num)-1;
+  if(all && param!=GENERATE_SIMILAR) tocheck = pow(2. , (double) num-1)-1;
+  else tocheck = pow(2. ,(double) num)-1;
   double step = tocheck/400;
 
   cout << "OBJECTIVE= 1 " << count << "/" << tocheck << endl;
